+++ lib_test/test_easy.ml
@@ -23,7 +23,9 @@ let test_iter () =
   assert_equal ~printer:pp_list
     ["ab"; "abab"] (Re.matches ~pos:2 ~len:7 re "abab ababab");
   assert_equal ~printer:pp_list
-    [""; ""] (Re.matches re_empty "ab");
+    [""; ""; ""] (Re.matches re_empty "ab");
+  assert_equal ~printer:pp_list
+    [""; "a"; ""] (Re.matches (Re.compile (Re.rep (Re.char 'a'))) "cat");
   ()
 
 let test_split () =
@@ -41,6 +43,32 @@ let test_split () =
     ["a "; "b"] (Re.split re_bow "a b");
   assert_equal ~printer:pp_list
     ["a"; " b"] (Re.split re_eow "a b");
+  assert_equal ~printer:pp_list
+    [] (Re.split re_whitespace "");
+  assert_equal ~printer:pp_list
+    [] (Re.split re_empty "");
+  ()
+
+let test_split_delim () =
+  assert_equal ~printer:pp_list
+    ["aa"; "bb"; "c"; "d"; ""] (Re.split_delim re_whitespace "aa bb c d ");
+  assert_equal ~printer:pp_list
+    ["a"; "b"; ""] (Re.split_delim ~pos:1 ~len:4 re_whitespace "aa b c d");
+  assert_equal ~printer:pp_list
+    [""; "a"; "full_word"; "bc"; ""]
+    (Re.split_delim re_whitespace " a full_word bc   ");
+  assert_equal ~printer:pp_list
+    [""; "a"; "b"; "c"; "d"; ""] (Re.split_delim re_empty "abcd");
+  assert_equal ~printer:pp_list
+    ["a"; "\nb"; ""] (Re.split_delim re_eol "a\nb");
+  assert_equal ~printer:pp_list
+    [""; "a "; "b"] (Re.split_delim re_bow "a b");
+  assert_equal ~printer:pp_list
+    ["a"; " b"; ""] (Re.split_delim re_eow "a b");
+  assert_equal ~printer:pp_list
+    [""] (Re.split_delim re_whitespace "");
+  assert_equal ~printer:pp_list
+    [""; ""] (Re.split_delim re_empty "");
   ()
 
 let map_split_delim =
@@ -69,8 +97,19 @@ let test_split_full () =
     [`D " "; `T "a"; `D " "; `T "full_word"; `D " "; `T "bc"; `D "   "]
     (Re.split_full re_whitespace " a full_word bc   " |> map_split_delim);
   assert_equal ~printer:pp_list'
-    [`D ""; `T "a"; `D ""; `T  "b"] (* XXX: not trivial *)
+    [] (Re.split_full re_whitespace "" |> map_split_delim);
+  assert_equal ~printer:pp_list'
+    [`D ""] (Re.split_full re_empty "" |> map_split_delim);
+  assert_equal ~printer:pp_list'
+    [`D ""; `T "a"; `D ""; `T  "b"; `D ""] (* XXX: not trivial *)
     (Re.split_full re_empty "ab" |> map_split_delim);
+  assert_equal ~printer:pp_list'
+    [] (Re.split_full re_whitespace "" |> map_split_delim);
+  assert_equal ~printer:pp_list'
+    [`D ""] (Re.split_full re_empty "" |> map_split_delim);
+  assert_equal ~printer:pp_list'
+    [`D ""; `T "c"; `D "a"; `T "t"; `D ""]
+    (Re.split_full (Re.compile (Re.rep (Re.char 'a'))) "cat" |> map_split_delim);
   ()
 
 let test_replace () =
@@ -80,6 +119,10 @@ let test_replace () =
     (Re.replace re ~f " hello world; I love chips!");
   assert_equal ~printer:pp_str " Allo maman, bobo"
     (Re.replace ~all:false re ~f " allo maman, bobo");
+  assert_equal ~printer:pp_str "a"
+    (Re.replace re_empty ~f:(fun _ -> "a") "");
+  assert_equal ~printer:pp_str "*c*t*"
+    (Re.replace (Re.compile (Re.rep (Re.char 'a'))) ~f:(fun _ -> "*") "cat");
   ()
 
 let test_replace_string () =
@@ -101,6 +144,7 @@ let test_bug_55 () =
 let suite = "easy" >:::
   [ "iter" >:: test_iter
   ; "split" >:: test_split
+  ; "split_delim" >:: test_split_delim
   ; "split_full" >:: test_split_full
   ; "replace" >:: test_replace
   ; "replace_string" >:: test_replace_string