+++ CHANGES.md
@@ -1,3 +1,8 @@
+Unreleased
+----------
+* Add `Re.split_delim` (#233)
+* Fix handling of empty matches in splitting and substitution functions (#233)
+
 1.11.0 (19-Aug-2023)
 --------------------